requirements.txt
gunicorn==22.0.0 ; python_version >= "3.9" \
    --hash=sha256:350679f91b24062c86e386e198a15438d53a7a8207235a78ba1b53df4c4378d9 \
    --hash=sha256:4a0b436239ff76fb33f11c07a16482c521a7e09c1ce3cc293c2330afe01bec63
packaging==23.2 ; python_version >= "3.9" \
    --hash=sha256:048fb0e9405036518eaaf48a55953c750c11e1a1b68e0dd1a9d62ed0c092cfc5 \
    --hash=sha256:8c491190033a9af7e1d931d0b5dacc2ef47509b34dd0de67ed209b5203fc88c7
scapy==2.4.5           # 用于网络包捕获和处理的库
requests==2.28.1       # 用于发送 HTTP 请求 (例如将数据包发送到远程服务器)
python-dotenv==1.0.0   # 用于从 .env 文件加载环境变量
Flask==2.1.3           # 用于构建 Web 控制台的轻量级 Web 框架